Move OrderMailerSubscriber#send_confirmation_email
We're moving towards having a separate subscriber for each type of
transaction email. This commit ensures that `OrderMailerSubscriber`
actions are all moved to subscriber classes that more accurately
describe the mailer action being taken.

Now, the old `OrderMailerSubscriber` does nothing and can be removed in
the next version of Solidus.

In subsequent commits we can now add additional subscribers for:

- Carton shipped transactional emails
- Short ship transactional emails
- Order cancellation transactionals emails
