+++ core/app/subscribers/spree/order_mailer_subscriber.rb
@@ -1,24 +1,13 @@
 # frozen_string_literal: true
 
 module Spree
-  # Mailing after events on a {Spree::Order}
   class OrderMailerSubscriber
     include Omnes::Subscriber
 
-    handle :order_finalized,
-           with: :send_confirmation_email,
-           id: :spree_order_mailer_send_confirmation_email
-
-    # Sends confirmation email to the user
-    #
-    # @param event [Omnes::UnstructuredEvent]
-    def send_confirmation_email(event)
-      order = event[:order]
-      unless order.confirmation_delivered?
-        Spree::Config.order_mailer_class.confirm_email(order).deliver_later
-        order.update_column(:confirmation_delivered, true)
-      end
-    end
+    def send_confirmation_email(_event) = nil
+    deprecate send_confirmation_email:
+      "Use Spree::OrderConfirmationMailerSubscriber#send_confirmation_email instead",
+      deprecator: Spree.deprecator
 
     def send_reimbursement_email(_event) = nil
     deprecate send_reimbursement_email:
