+++ core/lib/spree/core/engine.rb
@@ -71,7 +71,7 @@ module Spree
             reimbursement_errored
           ].each { |event_name| Spree::Bus.register(event_name) }
 
-          Spree::OrderMailerSubscriber.new.subscribe_to(Spree::Bus)
+          Spree::OrderConfirmationMailerSubscriber.new.subscribe_to(Spree::Bus)
           Spree::ReimbursementMailerSubscriber.new.subscribe_to(Spree::Bus)
         end
       end
